+++ smb2misc.c
@@ -8,6 +8,7 @@
 #include "nterr.h"
 #include "smb2pdu.h"
 #include "smb_common.h"
+#include "smbstatus.h"
 #include "mgmt/user_session.h"
 #include "connection.h"
