+++ smb2pdu.c
@@ -28,6 +28,7 @@
 #include "time_wrappers.h"
 #include "server.h"
 #include "smb_common.h"
+#include "smbstatus.h"
 #include "cifsd_work.h"
 #include "mgmt/user_config.h"
 #include "mgmt/share_config.h"
